+++ src/sampler/symbol_guess_linux.h
@@ -37,6 +37,7 @@ struct BuildStats {
   std::size_t string_labels = 0;
   std::size_t string_accumulated_labels = 0;
   std::size_t lambda_body_labels = 0;
+  std::size_t code_pattern_labels = 0;
   std::size_t thunk_candidates = 0;
   std::size_t thunk_resolved = 0;
   std::size_t thunk_labels = 0;